+++ index.js
@@ -1,29 +1,29 @@
 
-exports.status = require('./src/mojang/status')
-exports.getProfile = require('./src/mojang/get-profile')
-exports.getProfileHistory = require('./src/mojang/get-profile-history')
-exports.lookupProfiles = require('./src/mojang/lookup-profiles')
-exports.lookupProfileAt = require('./src/mojang/lookup-profile-at')
-exports.getOrdersStatistics = require('./src/mojang/get-orders-statistics')
+exports.status = require('./lib/mojang/status')
+exports.getProfile = require('./lib/mojang/get-profile')
+exports.getProfileHistory = require('./lib/mojang/get-profile-history')
+exports.lookupProfiles = require('./lib/mojang/lookup-profiles')
+exports.lookupProfileAt = require('./lib/mojang/lookup-profile-at')
+exports.getOrdersStatistics = require('./lib/mojang/get-orders-statistics')
 
-exports.authenticate = require('./src/yggdrasil/authenticate')
-exports.isValid = require('./src/yggdrasil/is-valid')
-exports.refresh = require('./src/yggdrasil/refresh')
-exports.invalidate = require('./src/yggdrasil/invalidate')
-exports.signout = require('./src/yggdrasil/signout')
+exports.authenticate = require('./lib/yggdrasil/authenticate')
+exports.isValid = require('./lib/yggdrasil/is-valid')
+exports.refresh = require('./lib/yggdrasil/refresh')
+exports.invalidate = require('./lib/yggdrasil/invalidate')
+exports.signout = require('./lib/yggdrasil/signout')
 
-exports.isSecure = require('./src/security/is-secure')
-exports.getChallenges = require('./src/security/get-challenges')
-exports.answerChallenges = require('./src/security/answer-challenges')
+exports.isSecure = require('./lib/security/is-secure')
+exports.getChallenges = require('./lib/security/get-challenges')
+exports.answerChallenges = require('./lib/security/answer-challenges')
 
-exports.getSession = require('./src/session/get-session')
-exports.getBlockedServers = require('./src/session/get-blocked-servers')
+exports.getSession = require('./lib/session/get-session')
+exports.getBlockedServers = require('./lib/session/get-blocked-servers')
 
-exports.getUser = require('./src/account/get-user')
-exports.setSkin = require('./src/account/set-skin')
-exports.resetSkin = require('./src/account/reset-skin')
-exports.getUserProfiles = require('./src/account/get-user-profiles')
-exports.getUserCapeData = require('./src/account/get-user-cape-data')
+exports.getUser = require('./lib/account/get-user')
+exports.setSkin = require('./lib/account/set-skin')
+exports.resetSkin = require('./lib/account/reset-skin')
+exports.getUserProfiles = require('./lib/account/get-user-profiles')
+exports.getUserCapeData = require('./lib/account/get-user-cape-data')
 
 // For document types, prefer dotless Array<Object>
 // https://github.com/jsdoc3/jsdoc/issues/1375
