+++ lib/mojang/get-orders-statistics.js
