+++ lib/mojang/get-profile-history.js
