+++ lib/mojang/lookup-profiles.js
