+++ lib/mojang/status.js
