+++ lib/on-api-error.js
