+++ lib/parse-res-bearer.js
