+++ lib/security/answer-challenges.js
