+++ lib/security/is-secure.js
