+++ lib/session/get-session.js
