+++ lib/yggdrasil/authenticate.js
