+++ lib/yggdrasil/invalidate.js
