+++ lib/yggdrasil/is-valid.js
