+++ lib/yggdrasil/signout.js
